Added authorization and authentication

+++ src/main/java/com/shekhar/projects/airBnbProject/Service/HotelServiceImp.java
@@ -5,13 +5,16 @@ import com.shekhar.projects.airBnbProject.Dto.HotelInfoDto;
 import com.shekhar.projects.airBnbProject.Dto.RoomDto;
 import com.shekhar.projects.airBnbProject.Entity.Hotel;
 import com.shekhar.projects.airBnbProject.Entity.Room;
+import com.shekhar.projects.airBnbProject.Entity.User;
 import com.shekhar.projects.airBnbProject.Repository.HotelRepository;
 import com.shekhar.projects.airBnbProject.Repository.RoomRepository;
 import com.shekhar.projects.airBnbProject.exception.ResourceNotFoundException;
+import com.shekhar.projects.airBnbProject.exception.UnAuthorisedException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -33,6 +36,10 @@ public class HotelServiceImp implements HotelService{
         log.info("Creating a new Hotel with name: {}",hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel,HotelDto.class);
@@ -44,6 +51,13 @@ public class HotelServiceImp implements HotelService{
         Hotel hotel =hotelRepository.findById(id).orElseThrow(()->
                 new ResourceNotFoundException("Hotel not find by id: "+id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
+
         return modelMapper.map(hotel,HotelDto.class);
     }
 
@@ -53,6 +67,10 @@ public class HotelServiceImp implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
         modelMapper.map(hotelDto,hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -65,6 +83,12 @@ public class HotelServiceImp implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         for(Room room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -78,6 +102,14 @@ public class HotelServiceImp implements HotelService{
         log.info("Activating the hotel with ID: {}",hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(()->
                 new ResourceNotFoundException("Hotel are not found with id: " +hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
+
         hotel.setActive(true);
 
         for(Room room: hotel.getRooms()) {
